Return a KeyEvent from the driver
