+++ vaitk/keys.py
@@ -409,3 +409,20 @@ def vai_key_code_to_text(key_code):
 
     return key_map.get(key_code, '')
 '''
+'''
+
+
+    @property
+    def text(self):
+        return vai_key_code_to_text(self.key_code)
+
+    @classmethod
+    def from_native_key_code(cls, native_key_code):
+        key_code = native_to_vai_key_code(native_key_code)
+        if key_code is None:
+            raise ValueError(f"Unknown native key code {native_key_code}")
+        return cls(key_code)
+
+
+
+'''
